# public/static/swagger-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Yeah! Time tracking API",
    "version": "1.0.0",
    "description": "Yeah! Time tracking API"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://localhost:3000"
    },
    {
      "url": "https://ytt-api.flbk.tech"
    }
  ],
  "components": {
    "securitySchemes": {
      "accessHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-access-token",
        "description": "Settings: API access token as header: Get token from `/sign-in` endpoint"
      },
      "accessBody": {
        "type": "apiKey",
        "in": "body",
        "name": "token",
        "description": "Settings: API access token as request body: Get token from `/sign-in` endpoint"
      },
      "accessQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "token",
        "description": "Settings: API access token as request URL query param: Get token from `/sign-in` endpoint"
      },
      "apiHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-token",
        "description": "API access token as header: Get token from `/authentication` endpoint"
      },
      "apiBody": {
        "type": "apiKey",
        "in": "body",
        "name": "apitoken",
        "description": "API access token as request body: Get token from `/authentication` endpoint"
      },
      "apiQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "apitoken",
        "description": "API access token as request URL query param: Get token from `/authentication` endpoint"
      }
    },
    "schemas": {
      "ApiErrorResponse": {
        "type": "object"
      },
      "Id": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "d741861a-3967-4503-90f2-afc8ee45fef8"
          }
        }
      },
      "UserId": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "example": "d741861a-3967-4503-90f2-afc8ee45fef8"
          }
        }
      },
      "Dates": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-03-06T09:13:23.522Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-03-06T09:13:23.522Z"
          }
        }
      },
      "AuthRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email address.",
            "example": "john@doe.test"
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "User password.",
            "example": "123456"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "User auth token.",
            "example": "dfdg.wertdf.wewetsfdfg"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          }
        }
      },
      "AuthUser": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Id"
          },
          {
            "$ref": "#/components/schemas/Dates"
          },
          {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "john"
              },
              "email": {
                "type": "string",
                "format": "email",
                "example": "john@doe.test"
              }
            }
          }
        ]
      },
      "ActionGroupRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "action group name.",
            "example": "action group #1"
          }
        }
      },
      "ActionGroupsResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/ActionGroup"
        }
      },
      "ActionGroupResponse": {
        "$ref": "#/components/schemas/ActionGroup"
      },
      "ActionGroup": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Id"
          },
          {
            "$ref": "#/components/schemas/UserId"
          },
          {
            "$ref": "#/components/schemas/Dates"
          },
          {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "action group #1"
              }
            }
          }
        ]
      },
      "ActionRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "action #1"
          },
          "type": {
            "type": "string",
            "example": "type #1"
          }
        }
      },
      "ActionsResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Action"
        }
      },
      "ActionResponse": {
        "$ref": "#/components/schemas/Action"
      },
      "Action": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Id"
          },
          {
            "$ref": "#/components/schemas/UserId"
          },
          {
            "$ref": "#/components/schemas/Dates"
          },
          {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "action #1"
              },
              "type": {
                "type": "string",
                "example": "type #1"
              },
              "actionGroupId": {
                "type": "string",
                "format": "uuid",
                "example": "d741861a-3967-4503-90f2-afc8ee45fef8"
              }
            }
          }
        ]
      }
    }
  },
  "paths": {
    "/auth/signin": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Failed response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Failed response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/settings/action-groups": {
      "get": {
        "tags": [
          "Settings - Action group"
        ],
        "security": [
          {
            "accessHeader": []
          },
          {
            "accessBody": []
          },
          {
            "accessQuery": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionGroupsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Failed response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Settings - Action group"
        ],
        "security": [
          {
            "accessHeader": []
          },
          {
            "accessBody": []
          },
          {
            "accessQuery": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionGroupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionGroupResponse"
                }
              }
            }
          },
          "400": {
            "description": "Failed response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/settings/action-groups/{id}": {
      "put": {
        "tags": [
          "Settings - Action group"
        ],
        "security": [
          {
            "accessHeader": []
          },
          {
            "accessBody": []
          },
          {
            "accessQuery": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "action group id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionGroupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionGroupResponse"
                }
              }
            }
          },
          "400": {
            "description": "Failed response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Settings - Action group"
        ],
        "security": [
          {
            "accessHeader": []
          },
          {
            "accessBody": []
          },
          {
            "accessQuery": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "action group id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionGroupResponse"
                }
              }
            }
          },
          "400": {
            "description": "Failed response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/settings/actions": {
      "get": {
        "tags": [
          "Settings - Action"
        ],
        "security": [
          {
            "accessHeader": []
          },
          {
            "accessBody": []
          },
          {
            "accessQuery": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionGroupsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Failed response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Settings - Action"
        ],
        "security": [
          {
            "accessHeader": []
          },
          {
            "accessBody": []
          },
          {
            "accessQuery": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Failed response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/settings/actions/{id}": {
      "put": {
        "tags": [
          "Settings - Action"
        ],
        "security": [
          {
            "accessHeader": []
          },
          {
            "accessBody": []
          },
          {
            "accessQuery": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "action id"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Failed response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Settings - Action"
        ],
        "security": [
          {
            "accessHeader": []
          },
          {
            "accessBody": []
          },
          {
            "accessQuery": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "action id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Failed response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}